More repos, services and exceptions.

+++ src/main/java/jact/lagaltproject/repositories/ChatRepository.java
@@ -3,5 +3,5 @@ package jact.lagaltproject.repositories;
 import jact.lagaltproject.models.Chat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChatRepository extends JpaRepository<Chat, Integer> {
+public interface ChatRepository extends JpaRepository<Chat, Long> {
 }
