+++ src/main/java/jact/lagaltproject/services/chat/ChatService.java
@@ -3,5 +3,5 @@ package jact.lagaltproject.services.chat;
 import jact.lagaltproject.models.Chat;
 import jact.lagaltproject.services.CrudService;
 
-public interface ChatService extends CrudService<Chat, Integer> {
+public interface ChatService extends CrudService<Chat, Long> {
 }
